requirements.txt
anyio==3.7.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
DateTime==5.1
dnspython==2.4.0
et-xmlfile==1.1.0
exceptiongroup==1.1.2
Flask==2.3.2
Flask-Session==0.5.0
fonttools==4.41.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
importlib-metadata==6.7.0
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
load-dotenv==0.1.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mysql-connector-python==8.0.33
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==1.5.2
Pillow==10.0.0
protobuf==3.20.3
pymongo==4.4.1
PyMySQL==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sniffio==1.3.0
soupsieve==2.4.1
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.15.0
zope.interface==6.0
